fix: 完成分页查询接口,前后端联调

+++ services/department-services/src/main/java/top/xblog1/emr/services/department/services/impl/DepartmentServicesImpl.java
@@ -8,6 +8,8 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
@@ -35,21 +37,25 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
-import static top.xblog1.emr.services.department.common.constant.RedisKeyConstant.DEPARTMENT_INFO_KEY_PREFIX;
+import static top.xblog1.emr.services.department.common.constant.RedisKeyConstant.*;
 import static top.xblog1.emr.services.department.common.enums.DepartmentCreateErrorCodeEnum.*;
 
 /**
- *
+ * 这张表很少更新，但是经常会被查询，尤其是全表查询
+ * 对于数据要永久存储到redis中，
+ * 采用版本号机制控制数据一致性，并设置版本号过期时间作为兜底
+ * 数据库更新只有管理员有权限，不需要上锁
  */
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class DepartmentServicesImpl implements DepartmentServices {
+    // todo 由于涉及缓存，需要加锁
 
     private final DepartmentMapper departmentMapper;
     private final AbstractChainContext abstractChainContext;
     private final DistributedCache distributedCache;
-    private final Mapper mapper;
+    private final RedissonClient redissonClient;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
@@ -59,16 +65,25 @@ public class DepartmentServicesImpl implements DepartmentServices {
         abstractChainContext.handler(DepartmentChainMarkEnum.DEPARTMENT_CREATE_FILTER.name(),requestParam);
         DepartmentDO departmentDO = BeanUtil.convert(requestParam, DepartmentDO.class);
         //log.info(departmentDO.toString());
+
+        RLock lock = redissonClient.getLock(LOCK_DEPARTMENT_INFO_KEY);
+        boolean tryLock = lock.tryLock();
+        if (!tryLock) {
+            throw new ServiceException("系统错误");
+        }
         try {
+            //插入数据
             departmentMapper.insert(departmentDO);
+            //删除版本号
+            instance.delete(DEPARTMENT_INFO_VERSION_CONTROLLER);
         }catch (DuplicateKeyException ex){
             throw new ClientException(HAS_NAME);
         } catch (Exception e) {
             log.error(e.getMessage());
             throw new ServiceException(DEPARTMENT_CREATE_FAIL);
+        }finally{
+            lock.unlock();
         }
-        instance.opsForValue().set(DEPARTMENT_INFO_KEY_PREFIX + departmentDO.getId(), JSONUtil.toJsonStr(departmentDO),30, TimeUnit.MINUTES);
-
         //log.info("部门{}创建完成",departmentDO);
         return departmentDO.getId();
     }
@@ -80,18 +95,25 @@ public class DepartmentServicesImpl implements DepartmentServices {
         if(id==null){
             throw new ClientException(ID_NOTNULL);
         }
+        RLock lock = redissonClient.getLock(LOCK_DEPARTMENT_INFO_KEY);
+        boolean tryLock = lock.tryLock();
+        if (!tryLock) {
+            throw new ServiceException("系统错误");
+        }
         try {
-            //先删除redis
+            //先数据库，再删缓存，因为数据库操作时间长，缓存操作时间短，这样可以尽量保证数据一致性
             instance.delete(DEPARTMENT_INFO_KEY_PREFIX+id);
             //再删除数据库 手动将更新时间置为null
             DepartmentDO departmentDO = departmentMapper.selectById(id);
             departmentDO.setUpdateTime(null);
             departmentMapper.updateById(departmentDO);
             departmentMapper.deleteById(id);
-            //再删除
-            instance.delete(DEPARTMENT_INFO_KEY_PREFIX+id);
+            //再删除版本号
+            instance.delete(DEPARTMENT_INFO_VERSION_CONTROLLER);
         } catch (Exception e) {
             throw new ServiceException(DEPARTMENT_DELETE_FAIL);
+        }finally {
+            lock.unlock();
         }
 
     }
@@ -112,15 +134,28 @@ public class DepartmentServicesImpl implements DepartmentServices {
         }
         //先更新数据库
         DepartmentDO departmentDO = BeanUtil.convert(requestParam, DepartmentDO.class);
-        int update = departmentMapper.updateById(departmentDO);
-        if(update == 0){
-            throw new ClientException(ID_NOTNULL);
+        RLock lock = redissonClient.getLock(LOCK_DEPARTMENT_INFO_KEY);
+        boolean tryLock = lock.tryLock();
+        if (!tryLock) {
+            throw new ServiceException("系统错误");
         }
-        //再删除缓存
-        StringRedisTemplate instance = (StringRedisTemplate) distributedCache.getInstance();
-        instance.delete(DEPARTMENT_INFO_KEY_PREFIX+departmentDO.getId());
-        DepartmentQueryRespDTO result = queryById(departmentDO.getId());
-        return BeanUtil.convert(result, DepartmentUpdateRespDTO.class);
+        try{
+
+            int update = departmentMapper.updateById(departmentDO);
+            if(update == 0){
+                throw new ClientException(ID_NOTNULL);
+            }
+            //再删除缓存
+            StringRedisTemplate instance = (StringRedisTemplate) distributedCache.getInstance();
+            instance.delete(DEPARTMENT_INFO_KEY_PREFIX+departmentDO.getId());
+            // 删除版本号
+            instance.delete(DEPARTMENT_INFO_VERSION_CONTROLLER);
+        }catch (DuplicateKeyException ex){
+            throw new ServiceException("更新失败");
+        }finally {
+            lock.unlock();
+        }
+        return BeanUtil.convert(departmentDO, DepartmentUpdateRespDTO.class);
     }
 
     @Override
@@ -140,15 +175,47 @@ public class DepartmentServicesImpl implements DepartmentServices {
         if(departmentDO==null){
             throw new ServiceException(CANNOT_FIND_DEPARTMENT);
         }
-        instance.opsForValue().set(DEPARTMENT_INFO_KEY_PREFIX + id, JSONUtil.toJsonStr(departmentDO),30, TimeUnit.MINUTES);
+        instance.opsForValue().set(DEPARTMENT_INFO_KEY_PREFIX + id, JSONUtil.toJsonStr(departmentDO));
         return BeanUtil.convert(departmentDO, DepartmentQueryRespDTO.class);
 
     }
 
     @Override
     public List<DepartmentQueryRespDTO> queryAll() {
+        //先检测计数器和全量缓存是否一致
         StringRedisTemplate instance = (StringRedisTemplate) distributedCache.getInstance();
+        //获取redis中的全量数据
         Set<String> keys = instance.keys(DEPARTMENT_INFO_KEY_PREFIX.concat("*"));
+        //获取版本号
+        String version = distributedCache.get(DEPARTMENT_INFO_VERSION_CONTROLLER, String.class);
+        //如果版本号与keys不一致或者版本号不存在
+        if(version==null||version.isEmpty()||!version.equals(String.valueOf(keys.size()))){
+            RLock lock = redissonClient.getLock(LOCK_DEPARTMENT_INFO_KEY);
+            boolean tryLock = lock.tryLock();
+            if (!tryLock) {
+                throw new ServiceException("系统错误");
+            }
+            List<DepartmentDO> departmentDOS;
+            try {
+                //查询数据库
+                departmentDOS = departmentMapper.selectList(Wrappers.lambdaQuery(DepartmentDO.class));
+                //批量删除
+                distributedCache.delete(keys);
+                //将数据库存入redis
+                departmentDOS.forEach(department -> {
+                    distributedCache.put(DEPARTMENT_INFO_KEY_PREFIX + department.getId(), JSONUtil.toJsonStr(department));
+                });
+                //把版本号填入redis，设置版本号过期时间，保证定期执行缓存更新操作，作为兜底策略
+                distributedCache.put(DEPARTMENT_INFO_VERSION_CONTROLLER, departmentDOS.size(), 30, TimeUnit.MINUTES);
+            }catch (DuplicateKeyException ex){
+                throw new ServiceException("系统执行错误");
+            }finally {
+                lock.unlock();
+            }
+             //返回值
+            return BeanUtil.convert(departmentDOS, DepartmentQueryRespDTO.class);
+        }
+        //如果版本号与keys数量一致，直接返回
         List<DepartmentDO> result = new ArrayList<>();
         keys.forEach(key -> {
             String s = instance.opsForValue().get(key);
